Button painting.
This commit adds the 'GuiPainterButton' function which will paint the
button at the current cursor position. The bounds of the button are
determined by the text size along with set horizontal and vertical padding
values.

This commit also modifies 'GuiPainterWindowBox' to not store the result
instead return on the function call.