Dropdown box painting.
This commit adds 'GuiPainterDropdownBox' to the API. A new struct
'GuiPainterDropdownBoxOptions' has been introduced that contains two
fields, 'active' and 'edit', which aids in managing the active and edit arguments
in the 'GuiDropdownBox' function. The width of the control is determined
by finding the largest string in the list. Button padding is also used for
this control.