Add 'GuiPainterTextBox' to the API.
This function is a wrapper for 'GuiTextBox'. It accepts a text pointer, the buffer
size, and a boolean pointer that updates the edit mode without requiring
the caller to modify the value. The width of the box is a constant set at 100.
An API will be introduced in the future to modify this width and to also take
into account the current width the cursor is operating in.