Add 'GuiPainterSameLine' function to API.
This function will move the cursor to the previous line passed the previous
control painted. This will allow multiple controls to be painted on the
same line. A new variable has been introduced to store the position
of the previous line. All API functions that advance the cursor will now
update the position of the previous line.