Add 'GuiPainterCheckBox' to the API.
This function is a wrapper for 'GuiCheckBox'. The check box size is determined
by the TEXT_SIZE gui style property along with the button padding. The box
will be displayed regardless if there is text or not. The 'checked' argument
is a pointer so that its state can be updated when clicked. The cursor will
also properly advance passed the box and text and include the TEXT_PADDING
property of the CHECKBOX control.

This function handles both raygui version 3.6 and 4.