Add 'GuiPainterValueBox' function to the API.
This function is a wrapper for 'GuiValueBox'. A new struct
'GuiPainterValueBoxOptions' has been introduced to pass all required
arguments in as a single argument. Text alignment is taken into account
when setting up the bounds for both rendering and cursor advancement.

The size of the value box uses a 9-digit string as the size. The edit mode is
also updated within this function.