Add 'GuiPainterSlider' function to the API.
This function is a wrapper for 'GuiSlider'. A new struct 'GuiPainterSliderOptions'
has been introduced to wrap multiple arguments to be passed in as one.
The 'value' field is updated when the slider value changes. The cursor
is properly advanced based on if text is present on both the left and right side
and padding is taken into account. The default size of the slider is 100. This
value will be mutable in a future API call.