Add 'GuiPainterListView' to the API.
This function is a wrapper for 'GuiListView'. A new struct
'GuiPainterListViewOptions' has been introduced that contains fields for
the scroll index, active index, and the requested number of visible items.
This number dictates how tall the list view will be. If it is 0, then a default
value of 5 is used. The width is determined by the largest text element
in the given list.